Camera and Resize (#146)
* Split back buffer into two buffers

* Add camera class

* Add work til now

* Changed Graph to use variables

* Changed some more values to variables

* Add scroll event

* Add resize event to ui_components

* Overhauled to use camera view matrix

* Add point buffer

* Modify mousehandler to use world coordinates

* Moving camera

* Overhauled camera again, just using translation and scale matrices

* Stay in bounds

* New reset coords

* Resizing world and final textures

* Fixed resize issue

* Add resize method to all ui components, and added a couple of special panels for controlpanel

* Fix CMakeLists

* Fix segfault on delete data

* Destroy textures

* Add initialization to ui/main

* Added larger world image

* fixed transparancy of panels

* Added scroll_margin to settings and camera_manager

* Fixed includes

* Removed unused graph variable

+++ src/ui/sdl/event/sdl_event_types.h
@@ -9,8 +9,8 @@
 
 class SDL_UIComponent;
 struct sdl_mouse_event_data : public mouse_event_data<SDL_UIComponent> {
-    uint8_t button;
-    sdl_mouse_event_data(uint32_t ty, uint8_t b,uint32_t time, vec2 &pos) : mouse_event_data<SDL_UIComponent>(ty, time, pos) {
+    int32_t button;
+    sdl_mouse_event_data(uint32_t ty, int32_t b,uint32_t time, vec2 &pos) : mouse_event_data<SDL_UIComponent>(ty, time, pos) {
         button = b;
     }
 };
